Factor out EncodingInterceptor, RedirectInterceptor, HttpFetchEnforcer from ApacheHttpFetcher
