+++ kifi-backend/modules/common/app/com/keepit/rover/fetcher/apache/ApacheHttpFetcher.scala
@@ -3,11 +3,8 @@ package com.keepit.rover.fetcher.apache
 import java.io.{ EOFException, IOException }
 import java.net._
 import java.security.GeneralSecurityException
-import java.security.cert.CertPathBuilderException
-import java.util.concurrent.atomic.{ AtomicInteger, AtomicReference }
 import java.util.concurrent.{ ConcurrentLinkedQueue, Executors, ThreadFactory, TimeUnit }
 import java.util.zip.ZipException
-import javax.net.ssl.{ SSLException, SSLHandshakeException }
 
 import com.keepit.common.akka.SafeFuture
 import com.keepit.common.concurrent.ExecutionContext
@@ -15,7 +12,6 @@ import com.keepit.common.service.RequestConsolidator
 import com.keepit.common.strings._
 import com.keepit.rover.fetcher._
 import com.keepit.scraper.DeprecatedHttpInputStream
-import play.api.Play.current
 
 import com.keepit.common.time._
 import com.keepit.common.healthcheck.AirbrakeNotifier
@@ -26,36 +22,40 @@ import com.keepit.common.plugin.SchedulingProperties
 import com.keepit.model.HttpProxy
 import com.keepit.scraper._
 import org.apache.http.HttpHeaders._
-import org.apache.http.HttpStatus._
 import org.apache.http._
 import org.apache.http.auth.{ AuthScope, UsernamePasswordCredentials }
 import org.apache.http.client.config.RequestConfig
-import org.apache.http.client.entity.{ DeflateDecompressingEntity, GzipDecompressingEntity }
 import org.apache.http.client.methods.{ CloseableHttpResponse, HttpGet }
 import org.apache.http.client.protocol.HttpClientContext
 import org.apache.http.config.RegistryBuilder
 import org.apache.http.conn.socket.{ ConnectionSocketFactory, PlainConnectionSocketFactory }
-import org.apache.http.conn.{ ConnectTimeoutException, HttpHostConnectException }
 import org.apache.http.impl.client.{ BasicCredentialsProvider, HttpClientBuilder }
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
 import org.apache.http.protocol.{ HttpContext, BasicHttpContext }
 import org.apache.http.util.EntityUtils
 import org.joda.time.DateTime
-import play.api.Play
-import sun.security.validator.ValidatorException
-import views.html.helper.input
+import RedirectInterceptor._
 
 import scala.concurrent.Future
 import scala.ref.WeakReference
 import scala.util.{ Failure, Success, Try }
 import scala.concurrent.duration._
 
-import org.apache.commons.io.{ IOUtils, FileUtils }
+import org.apache.commons.io.IOUtils
+
+import com.keepit.rover.fetcher.{ DeprecatedFetcherHttpContext, HttpRedirect }
+import org.apache.http.protocol.{ HttpContext => ApacheHttpContext }
+
+object ApacheHttpFetcher {
+  private[apache] implicit class DeprecatedFetcherHttpContextAdaptor(context: ApacheHttpContext) extends DeprecatedFetcherHttpContext {
+    def destinationUrl: Option[String] = Option(context.getAttribute(scraperDestinationUrlAttribute).asInstanceOf[String])
+    def redirects: Seq[HttpRedirect] = Option(context.getAttribute(redirectsAttribute).asInstanceOf[Seq[HttpRedirect]]).getOrElse(Seq.empty[HttpRedirect])
+  }
+}
 
 // based on Apache HTTP Client (this one is blocking but feature-rich & flexible; see http://hc.apache.org/httpcomponents-client-ga/index.html)
 class ApacheHttpFetcher(val airbrake: AirbrakeNotifier, userAgent: String, connectionTimeout: Int, soTimeOut: Int, schedulingProperties: SchedulingProperties, scraperHttpConfig: ScraperHttpConfig) extends DeprecatedHttpFetcher with Logging with ScraperUtils {
-
-  implicit def toFetcherContext(apacheCtx: HttpContext): DeprecatedFetcherHttpContext = new DeprecatedFetcherHttpContextAdaptor(apacheCtx)
+  import ApacheHttpFetcher._
 
   val cm = {
     val registry = RegistryBuilder.create[ConnectionSocketFactory]
@@ -66,77 +66,15 @@ class ApacheHttpFetcher(val airbrake: AirbrakeNotifier, userAgent: String, conne
   cm.setMaxTotal(100)
 
   val defaultRequestConfig = RequestConfig.custom().setConnectTimeout(connectionTimeout).setSocketTimeout(soTimeOut).build()
-
-  private val httpClientBuilder = HttpClientBuilder.create()
-  httpClientBuilder.setDefaultRequestConfig(defaultRequestConfig)
-  httpClientBuilder.setConnectionManager(cm)
-  httpClientBuilder.setUserAgent(userAgent)
-  httpClientBuilder.setRequestExecutor(new ContentAwareHttpRequestExecutor())
-
-  // track redirects
-  val redirectInterceptor = new HttpResponseInterceptor() {
-    override def process(response: HttpResponse, context: HttpContext) {
-      if (response.containsHeader(LOCATION)) {
-        val locations = response.getHeaders(LOCATION)
-        if (locations.length > 0) {
-          val currentLocation = context.getAttribute("scraper_destination_url").asInstanceOf[String]
-          val redirect = HttpRedirect.withStandardizationEffort(response.getStatusLine.getStatusCode, currentLocation, locations(0).getValue())
-          val redirects = context.getAttribute("redirects").asInstanceOf[Seq[HttpRedirect]] :+ redirect
-          context.setAttribute("redirects", redirects)
-          context.setAttribute("scraper_destination_url", redirect.newDestination)
-        }
-      }
-    }
-  }
-  httpClientBuilder.addInterceptorFirst(redirectInterceptor)
-
-  // transfer encoding
-  val encodingInterceptor = new HttpResponseInterceptor() {
-    override def process(response: HttpResponse, context: HttpContext) {
-      val entity = response.getEntity()
-      if (entity != null) {
-        val ceheader = entity.getContentEncoding()
-        if (ceheader != null) {
-          val codecs = ceheader.getElements()
-          codecs.foreach { codec =>
-            if (codec.getName().equalsIgnoreCase("gzip")) {
-              response.setEntity(new GzipDecompressingEntity(response.getEntity()))
-              return
-            }
-            if (codec.getName().equalsIgnoreCase("deflate")) {
-              response.setEntity(new DeflateDecompressingEntity(response.getEntity()))
-              return
-            }
-          }
-          val encoding = codecs.map(_.getName).mkString(",")
-          log.error(s"unsupported content-encoding: ${encoding}")
-        }
-      }
-    }
-  }
-  httpClientBuilder.addInterceptorFirst(encodingInterceptor)
-
-  val httpClient = httpClientBuilder.build()
-
-  val LONG_RUNNING_THRESHOLD = if (Play.maybeApplication.isDefined && Play.isDev) 1000 else sys.props.get("fetcher.abort.threshold") map (_.toInt) getOrElse (2 * 1000 * 60) // Play reference can be removed
-  val Q_SIZE_THRESHOLD = scraperHttpConfig.httpFetcherQSizeThreshold
-
-  case class FetchInfo(url: String, ts: Long, htpGet: HttpGet, thread: Thread) {
-    val killCount = new AtomicInteger()
-    val respStatusRef = new AtomicReference[StatusLine]()
-    val exRef = new AtomicReference[Throwable]()
-    override def toString = s"[Fetch($url,${ts},${thread.getName})] isAborted=${htpGet.isAborted} killCount=${killCount} respRef=${respStatusRef} exRef=${exRef}"
-  }
-
-  private def removeRef(iter: java.util.Iterator[_], msgOpt: Option[String] = None) {
-    try {
-      for (msg <- msgOpt)
-        log.info(msg)
-      iter.remove()
-    } catch {
-      case t: Throwable =>
-        log.error(s"[terminator] Caught exception $t; (cause=${t.getCause}) while attempting to remove entry from queue")
-    }
+  val httpClient = {
+    val httpClientBuilder = HttpClientBuilder.create()
+    httpClientBuilder.setDefaultRequestConfig(defaultRequestConfig)
+    httpClientBuilder.setConnectionManager(cm)
+    httpClientBuilder.setUserAgent(userAgent)
+    httpClientBuilder.setRequestExecutor(new ContentAwareHttpRequestExecutor())
+    httpClientBuilder.addInterceptorFirst(new RedirectInterceptor()) // track redirects
+    httpClientBuilder.addInterceptorFirst(new EncodingInterceptor()) // transfer encoding
+    httpClientBuilder.build()
   }
 
   private def logAndSet[T](fetchInfo: FetchInfo, ret: T)(t: Throwable, tag: String, ctx: String, notify: Boolean = false): T = {
@@ -146,76 +84,21 @@ class ApacheHttpFetcher(val airbrake: AirbrakeNotifier, userAgent: String, conne
   }
 
   val q = new ConcurrentLinkedQueue[WeakReference[FetchInfo]]()
-  val enforcer = new Runnable {
-    def run(): Unit = {
-      try {
-        log.debug(s"[enforcer] checking for long running fetch requests ... q.size=${q.size}")
-        if (!q.isEmpty) {
-          val iter = q.iterator
-          while (iter.hasNext) {
-            val curr = System.currentTimeMillis
-            val ref = iter.next
-            ref.get map {
-              case ft: FetchInfo =>
-                if (ft.respStatusRef.get != null) {
-                  val sc = ft.respStatusRef.get.getStatusCode
-                  removeRef(iter, if (sc != SC_OK && sc != SC_NOT_MODIFIED) Some(s"[enforcer] ${ft.url} finished with abnormal status:${ft.respStatusRef.get}") else None)
-                } else if (ft.exRef.get != null) removeRef(iter, Some(s"[enforcer] ${ft.url} caught error ${ft.exRef.get}; remove from q"))
-                else if (ft.htpGet.isAborted) removeRef(iter, Some(s"[enforcer] ${ft.url} is aborted; remove from q"))
-                else {
-                  val runMillis = curr - ft.ts
-                  if (runMillis > LONG_RUNNING_THRESHOLD * 2) {
-                    val msg = s"[enforcer] attempt# ${ft.killCount.get} to abort long ($runMillis ms) fetch task: ${ft.htpGet.getURI}"
-                    log.warn(msg)
-                    ft.htpGet.abort() // inform scraper
-                    ft.killCount.incrementAndGet()
-                    log.debug(s"[enforcer] ${ft.htpGet.getURI} isAborted=${ft.htpGet.isAborted}")
-                    if (!ft.htpGet.isAborted) {
-                      log.warn(s"[enforcer] failed to abort long ($runMillis ms) fetch task $ft; calling interrupt ...")
-                      ft.thread.interrupt
-                      if (ft.thread.isInterrupted) {
-                        log.warn(s"[enforcer] thread ${ft.thread} has been interrupted for fetch task $ft")
-                        // removeRef -- maybe later
-                      } else {
-                        val msg = s"[enforcer] attempt# ${ft.killCount.get} failed to interrupt ${ft.thread} for fetch task $ft"
-                        log.error(msg)
-                        if (ft.killCount.get % 5 == 0)
-                          airbrake.notify(msg)
-                      }
-                    }
-                  } else if (runMillis > LONG_RUNNING_THRESHOLD) {
-                    log.warn(s"[enforcer] potential long ($runMillis ms) running task: $ft; stackTrace=${ft.thread.getStackTrace.mkString("|")}")
-                  } else {
-                    log.debug(s"[enforcer] $ft has been running for $runMillis ms")
-                  }
-                }
-            } orElse {
-              removeRef(iter)
-              None
-            }
-          }
-        }
-        if (q.size > Q_SIZE_THRESHOLD) {
-          airbrake.notify(s"[enforcer] q.size (${q.size}) crossed threshold ($Q_SIZE_THRESHOLD)")
-        } else if (q.size > Q_SIZE_THRESHOLD / 2) {
-          log.warn(s"[enforcer] q.size (${q.size}) crossed threshold/2 ($Q_SIZE_THRESHOLD)")
-        }
-      } catch {
-        case t: Throwable =>
-          airbrake.notify(s"[enforcer] Caught exception $t; queue=$q; cause=${t.getCause}; stack=${t.getStackTraceString}")
-      }
-    }
-  }
-  val scheduler = Executors.newSingleThreadScheduledExecutor(new ThreadFactory {
-    def newThread(r: Runnable): Thread = {
-      val thread = new Thread(r, "HttpFetcher-Enforcer")
-      log.debug(s"[HttpFetcher] $thread created")
-      thread
-    }
-  })
-
-  val ENFORCER_FREQ: Int = scraperHttpConfig.httpFetcherEnforcerFreq
   if (schedulingProperties.enabled) {
+    val enforcer = {
+      val Q_SIZE_THRESHOLD = scraperHttpConfig.httpFetcherQSizeThreshold
+      new HttpFetchEnforcer(q, Q_SIZE_THRESHOLD, airbrake)
+    }
+
+    val scheduler = Executors.newSingleThreadScheduledExecutor(new ThreadFactory {
+      def newThread(r: Runnable): Thread = {
+        val thread = new Thread(r, "HttpFetcher-Enforcer")
+        log.debug(s"[HttpFetcher] $thread created")
+        thread
+      }
+    })
+
+    val ENFORCER_FREQ: Int = scraperHttpConfig.httpFetcherEnforcerFreq
     scheduler.scheduleWithFixedDelay(enforcer, ENFORCER_FREQ, ENFORCER_FREQ, TimeUnit.SECONDS)
   }
